Moved Theme Config to Body folder.
